Rename back to ParsePhonePushSample for easier maintenance
