+++ WindowsPhone/ParsePhonePushSample/Assets/ApplicationIcon.png
