+++ WindowsPhone/ParsePhonePushSample/Assets/Tiles/FlipCycleTileLarge.png
