+++ WindowsPhone/ParsePhonePushSample/Assets/Tiles/FlipCycleTileMedium.png
