+++ WindowsPhone/ParsePhonePushSample/Assets/Tiles/FlipCycleTileSmall.png
