+++ WindowsPhone/ParsePhonePushSample/Assets/Tiles/IconicTileMediumLarge.png
