+++ WindowsPhone/ParsePhonePushSample/Assets/Tiles/IconicTileSmall.png
