+++ WindowsPhone/ParsePhonePushSample/MainPage.xaml
@@ -1,5 +1,5 @@
 ﻿<phone:PhoneApplicationPage
-    x:Class="ParsePhonePushTutorial.MainPage"
+    x:Class="ParsePhonePushSample.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:phone="clr-namespace:Microsoft.Phone.Controls;assembly=Microsoft.Phone"
