+++ WindowsPhone/ParsePhonePushSample/Properties/AppManifest.xml
