+++ WindowsPhone/ParsePhonePushSample/Properties/AssemblyInfo.cs
@@ -6,11 +6,11 @@ using System.Resources;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("ParsePhonePushTutorial")]
+[assembly: AssemblyTitle("ParsePhonePushSample")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("ParsePhonePushTutorial")]
+[assembly: AssemblyProduct("ParsePhonePushSample")]
 [assembly: AssemblyCopyright("Copyright ©  2013")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
