+++ WindowsPhone/ParsePhonePushSample/Resources/AppResources.resx
