+++ WindowsPhone/ParsePhonePushSample/lib/Parse.Phone.xml
