+++ WindowsPhone/ParsePhonePushSample/lib/Parse.xml
