--- WindowsPhone/ParsePhonePushTutorial.userprefs
@@ -1,8 +1,0 @@
-﻿<Properties StartupItem="ParsePhonePushSample/ParsePhonePushSample.csproj">
-  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
-  <MonoDevelop.Ide.Workbench />
-  <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore />
-  </MonoDevelop.Ide.DebuggingService.Breakpoints>
-  <MonoDevelop.Ide.DebuggingService.PinnedWatches />
-</Properties>
